// src/app/model/client.model.ts
export class Client {
    _id: String;
    username: string;
    password: string;
    firstName: string;
    lastName: string;
    phone: String;
    agency: String;
    token?: string;
    cin: string;
    cinDate: Date;
    agenceId: String;
    birthday: String;
    newPassword: String;
    oldPassword: String;

    name: string;
    surname: string;
    address: string;
    email: String;

    postalCode: String;
    drivingLicenceType: String;
    state: String;
    drivingLicenceNum: String;
    hasPack: String;
}